requirements.txt
Django==5.1.6
Pillow==11.2.1
